Data.Generic.HKD.Types

Description

Documentation

newtype HKD (structure :: Type) (f :: Type -> Type)

Higher-kinded data (HKD) is the design pattern in which every field in our type is wrapped in some functor f:

  data User f
   = User
       { name :: f String
       , age  :: f Int
       }

Depending on the functor, we can get different behaviours: with Maybe, we get a partial structure; with Validation, we get a piecemeal validator; and so on. The HKD newtype allows us to lift any type into an HKD-style API via its generic representation.

>>> :set -XDeriveGeneric -XTypeApplications
>>> :{
data User
  = User { name :: String, age :: Int }
  deriving Generic
:}

The HKD type is indexed by our choice of functor and the structure we're lifting. In other words, we can define a synonym for our behaviour:

>>> import Data.Monoid (Last (..))
>>> type Partial a = HKD a Last

... and then we're ready to go!

>>> mempty @(Partial User)
User {name = Last {getLast = Nothing}, age = Last {getLast = Nothing}}

>>> mempty @(HKD (Int, Bool) [])
(,) [] []

Constructors

HKD

Fields

runHKD :: HKD_ f structure Void

type HKD_ (f :: Type -> Type) (structure :: Type) = GHKD_ f (Rep structure)

Calculate the "partial representation" of a type.

type family GHKD_ (f :: Type -> Type) (rep :: Type -> Type) = (output :: Type -> Type) | output -> f rep where ...

Calculate the "partial representation" of a generic rep.

Equations

GHKD_ f (M1 index meta inner) = M1 index meta (GHKD_ f inner)

GHKD_ f (left :*: right) = GHKD_ f left :*: GHKD_ f right

GHKD_ f (K1 index value) = K1 index (f value)

GHKD_ f (left :+: right) = GHKD_ f left :+: GHKD_ f right

class Tuple (f :: Type -> Type) (structure :: Type) (tuple :: Type) | f structure -> tuple where

Methods

toTuple :: HKD structure f -> tuple

fromTuple :: tuple -> HKD structure f